refactor: renamed provider name; replace context url
