+++ packages/example-documents/src/examples/UniversityDegreeCredentialNewRenderer.json
@@ -2,7 +2,7 @@
   "@context": [
     "https://www.w3.org/ns/credentials/v2",
     "https://www.w3.org/ns/credentials/examples/v2",
-    "http://localhost:4000/render-template-context-2024.json"
+    "https://vckit-contexts.s3.ap-southeast-2.amazonaws.com/render-template-context-2024.json"
   ],
   "id": "http://example.edu/credentials/3732",
   "type": ["VerifiableCredential", "UniversityDegreeCredential"],
