+++ packages/renderer/src/index.ts
@@ -1,6 +1,6 @@
 import { Renderer } from './renderer.js';
 import { WebRenderingTemplate2022 } from './providers/web-rendering-template-2022.js';
-import { RenderTemplate2024 } from './providers/web-rendering-template-2024.js';
+import { RenderTemplate2024 } from './providers/render-template-2024.js';
 
 export { Renderer, WebRenderingTemplate2022, RenderTemplate2024 };
 export { RenderDefaultContexts } from './render-default-contexts.js';
