+++ packages/renderer/src/providers/render-template-2024.ts
